refactor: streamline RichTextEditor components and enhance toolbar functionality
